shell: command measuring CPU usage
This is a copy of musiclover's function:
https://forum.chibios.org/viewtopic.php?p=29448#p29448
Note, that it only deals with run time accumulated since system start, so it is
not suited for real-time measurement. Its purpose is to roughly determine the
state of the system in a longer run, e.g. after adding some new cpu-intensive
code.